+++ client.go
@@ -1,7 +1,12 @@
 package http
 
 import (
+	"crypto/tls"
+	"crypto/x509"
 	"fmt"
+	"net/url"
+	"runtime"
+	"strings"
 	"time"
 )
 
@@ -23,15 +28,73 @@ type Client interface {
 type ClientConfiguration struct {
 	// URL is the base URL for requests.
 	URL string `json:"url" yaml:"url" comment:"Base URL of the server to connect."`
-	// CaCert is either the CA certificate to expect on the server in PEM format
+	// CACert is either the CA certificate to expect on the server in PEM format
 	//         or the name of a file containing the PEM.
-	CaCert string `json:"cacert" yaml:"cacert" comment:"CA certificate in PEM format to use for host verification. Note: due to a bug in Go on Windows this has to be explicitly provided."`
+	CACert string `json:"cacert" yaml:"cacert" comment:"CA certificate in PEM format to use for host verification. Note: due to a bug in Go on Windows this has to be explicitly provided."`
 	// Timeout is the time the client should wait for a response.
 	Timeout time.Duration `json:"timeout" yaml:"timeout" comment:"HTTP call timeout." default:"2s"`
 	// ClientCert is a PEM containing an x509 certificate to present to the server or a file name containing the PEM.
 	ClientCert string `json:"cert" yaml:"cert" comment:"Client certificate file in PEM format."`
 	// ClientKey is a PEM containing a private key to use to connect the server or a file name containing the PEM.
 	ClientKey string `json:"key" yaml:"key" comment:"Client key file in PEM format."`
+
+	// caCertPool is for internal use only. It contains the loaded CA certificates after Validate.
+	caCertPool *x509.CertPool
+	// cert is for internal use only. It contains the loaded TLS key and certificate after Validate.
+	cert *tls.Certificate
+}
+
+// Validate validates the client configuration and returns an error if it is invalid.
+func (c *ClientConfiguration) Validate() error {
+	_, err := url.ParseRequestURI(c.URL)
+	if err != nil {
+		return fmt.Errorf("invalid URL: %s", c.URL)
+	}
+	if c.Timeout < 100*time.Millisecond {
+		return fmt.Errorf("timeout value %s is too low, must be at least 100ms", c.Timeout.String())
+	}
+
+	if strings.TrimSpace(c.CACert) != "" {
+		caCert, err := loadPem(c.CACert)
+		if err != nil {
+			return fmt.Errorf("failed to load CA certificate (%w)", err)
+		}
+
+		c.caCertPool = x509.NewCertPool()
+		if !c.caCertPool.AppendCertsFromPEM(caCert) {
+			return fmt.Errorf("invalid CA certificate provided")
+		}
+	} else if runtime.GOOS == "windows" && strings.HasPrefix(c.URL, "https://") {
+		//Remove if https://github.com/golang/go/issues/16736 gets fixed
+		return fmt.Errorf(
+			"no CA certificate provided for HTTPS query while running on Windows: due to a bug (#16736) in " +
+				"Golang on Windows CA certificates have to be explicitly provided for https:// URLs",
+		)
+	}
+
+	if c.ClientCert != "" && c.ClientKey == "" {
+		return fmt.Errorf("client certificate provided without client key")
+	} else if c.ClientCert == "" && c.ClientKey != "" {
+		return fmt.Errorf("client key provided without client certificate")
+	}
+
+	if c.ClientCert != "" && c.ClientKey != "" {
+		clientCert, err := loadPem(c.ClientCert)
+		if err != nil {
+			return fmt.Errorf("failed to load client certificate (%w)", err)
+		}
+		clientKey, err := loadPem(c.ClientKey)
+		if err != nil {
+			return fmt.Errorf("failed to load client certificate (%w)", err)
+		}
+		cert, err := tls.X509KeyPair(clientCert, clientKey)
+		if err != nil {
+			return fmt.Errorf("failed to load certificate or key (%w)", err)
+		}
+		c.cert = &cert
+	}
+
+	return nil
 }
 
 // FailureReason describes the Reason why the request failed.