+++ client_factory.go
@@ -2,10 +2,7 @@ package http
 
 import (
 	"crypto/tls"
-	"crypto/x509"
-	"fmt"
 	"net/http"
-	"runtime"
 	"strings"
 
 	"github.com/containerssh/log"
@@ -16,7 +13,14 @@ func NewClient(
 	config ClientConfiguration,
 	logger log.Logger,
 ) (Client, error) {
-	tlsConfig, err := createTLSConfig(config, logger)
+	if err := config.Validate(); err != nil {
+		return nil, err
+	}
+	if logger == nil {
+		panic("BUG: no logger provided for http.NewClient")
+	}
+
+	tlsConfig, err := createTLSConfig(config)
 	if err != nil {
 		return nil, err
 	}
@@ -38,15 +42,8 @@ func NewClient(
 	}, nil
 }
 
-func createTLSConfig(config ClientConfiguration, logger log.Logger) (*tls.Config, error) {
-	if logger == nil {
-		return nil, fmt.Errorf("no logger provided")
-	}
-
-	if config.URL == "" {
-		return nil, fmt.Errorf("no URL provided")
-	}
-
+// createTLSConfig creates a TLS config. Should only be called after config.Validate().
+func createTLSConfig(config ClientConfiguration) (*tls.Config, error) {
 	if !strings.HasPrefix(config.URL, "https://") {
 		return nil, nil
 	}
@@ -62,37 +59,11 @@ func createTLSConfig(config ClientConfiguration, logger log.Logger) (*tls.Config
 			tls.TLS_CHACHA20_POLY1305_SHA256,
 		},
 	}
-	if strings.TrimSpace(config.CaCert) != "" {
-		caCert, err := loadPem(config.CaCert)
-		if err != nil {
-			return nil, fmt.Errorf("failed to load CA certificate (%w)", err)
-		}
-
-		caCertPool := x509.NewCertPool()
-		caCertPool.AppendCertsFromPEM(caCert)
-		tlsConfig.RootCAs = caCertPool
-	} else if runtime.GOOS == "windows" && strings.HasPrefix(config.URL, "https://") {
-		//Remove if https://github.com/golang/go/issues/16736 gets fixed
-		return nil, fmt.Errorf(
-			"due to a bug (#16736) in Golang on Windows CA certificates have to be explicitly " +
-				"provided for https:// authentication server URLs",
-		)
+	if config.caCertPool != nil {
+		tlsConfig.RootCAs = config.caCertPool
 	}
-
-	if config.ClientCert != "" && config.ClientKey != "" {
-		clientCert, err := loadPem(config.ClientCert)
-		if err != nil {
-			return nil, fmt.Errorf("failed to load client certificate (%w)", err)
-		}
-		clientKey, err := loadPem(config.ClientKey)
-		if err != nil {
-			return nil, fmt.Errorf("failed to load client certificate (%w)", err)
-		}
-		cert, err := tls.X509KeyPair(clientCert, clientKey)
-		if err != nil {
-			return nil, fmt.Errorf("failed to load certificate or key (%w)", err)
-		}
-		tlsConfig.Certificates = []tls.Certificate{cert}
+	if config.cert != nil {
+		tlsConfig.Certificates = []tls.Certificate{*config.cert}
 	}
 	return tlsConfig, nil
 }